109 handle warnings (#114)
* Fix C4265 Missing virtual destructor

* Solve C4062 no default case in switch

* Solves C4464 path includes ".." for Vizir

* No more extern library warning

* Fix CMake MSVC linking and CPP version errors

* Clean CMakeLists.txt

* Fix MSVC noexcept warning

* Fix C4715

* Fix type implicit conversion warning

* Fix MSVC warning bug

* Fix Geotrup missing Bob linking

---------

+++ Caduq/src/Caduq/Objects/Spline.cpp
@@ -154,11 +154,11 @@ namespace Caduq
     void Spline::SetPopupParam(EntityManager& entityManager, Geometry::SplinePoint startPoint, int startPointID, Geometry::SplinePoint endPoint, int endPointID)
     {
         for (int i = 0; i < 3; i++)
-            m_GuiStartTangent[i] = startPoint.tangent[i];
+            m_GuiStartTangent[i] = static_cast<float>(startPoint.tangent[i]);
         m_GuiStartTension[0] = static_cast<float>(startPoint.tension);
 
         for (int i = 0; i < 3; i++)
-            m_GuiEndTangent[i] = endPoint.tangent[i];
+            m_GuiEndTangent[i] = static_cast<float>(endPoint.tangent[i]);
         m_GuiEndTension[0] = static_cast<float>(endPoint.tension);
 
         for (int i = 0; i < entityManager.GetPointList().size(); i++)
